Table design revamp //WIP
