+++ src/components/App/FarmList.tsx
@@ -26,19 +26,10 @@ interface Props {
 
 const FarmsList: FC<Props> = ({ farms, positions }) => {
   const router = useRouter();
-
-  useEffect(() => {
-    console.log("positions in farmlist", positions);
-  }, [positions]);
-
   return (
     <>
       {farms.map((farm: FarmType) => {
         const tokenNames = formatTokenSymbols(farm?.asset.symbol);
-        // const tokenSymbol =
-        //   tokenNames.length == 2
-        //     ? `${tokenNames[0]}-${tokenNames[1]}`
-        //     : tokenNames[0];
         const safetyScore = (farm?.safetyScore * 10).toFixed(1);
         const position =
           positions[
@@ -48,11 +39,11 @@ const FarmsList: FC<Props> = ({ farms, positions }) => {
           position?.unstaked.amountUSD + position?.staked.amountUSD;
 
         return (
-          <tr key={`${farm.asset.address}-${farm.tvl}`} className="group">
+          <tr key={`${farm.asset.address}-${farm.tvl}`}>
             <td className="whitespace-nowrap max-w-[288px] py-8 text-sm pl-8 md:pl-14 lg:pl-28">
               <div className="flex flex-col gap-y-[10px]">
                 <div className="flex flex-row items-center">
-                  <div className="text-blueSilver font-bold text-base leading-5">
+                  <div className="font-bold text-base leading-5">
                     {tokenNames.map((tokenName, index) => (
                       <span key={index} className="mr-[3px]">
                         {tokenName}
@@ -61,7 +52,7 @@ const FarmsList: FC<Props> = ({ farms, positions }) => {
                     ))}
                   </div>
                 </div>
-                <div className="text-[#9397A6] font-medium text-base leading-5">
+                <div className="font-medium text-base leading-5">
                   {formatFirstLetter(farm?.protocol)} on{" "}
                   {formatFirstLetter(farm?.chain)}
                 </div>
@@ -71,10 +62,10 @@ const FarmsList: FC<Props> = ({ farms, positions }) => {
             <td className="hidden lg:table-cell whitespace-nowrap">
               <FarmAssets logos={farm?.asset.logos} />
             </td>
-            <td className="whitespace-nowrap py-8 text-right sm:pr-3 sm:pl-4 text-blueSilver font-bold text-base leading-5 tracking-wide">
+            <td className="whitespace-nowrap py-8 text-right sm:pr-3 sm:pl-4 font-bold text-base leading-5 tracking-wide">
               {toDollarUnits(farm?.tvl)}
             </td>
-            <td className="whitespace-nowrap py-8 pl-0 pr-2 text-blueSilver font-bold text-base leading-5 tracking-wide">
+            <td className="whitespace-nowrap py-8 pl-0 pr-2 font-bold text-base leading-5 tracking-wide">
               <div className="w-full inline-flex justify-end items-center gap-x-2">
                 <Tooltip
                   label={
@@ -101,10 +92,10 @@ const FarmsList: FC<Props> = ({ farms, positions }) => {
                 </Tooltip>
               </div>
             </td>
-            <td className="hidden md:table-cell whitespace-nowrap max-w-[130px] h-full py-0 pl-0 lg:pl-16 pr-3 text-blueSilver font-bold text-base leading-5 tracking-wide">
+            <td className="hidden md:table-cell whitespace-nowrap max-w-[130px] h-full py-0 pl-0 lg:pl-16 pr-3 font-bold text-base leading-5 tracking-wide">
               <Rewards rewards={farm?.rewards} />
             </td>
-            <td className="hidden lg:table-cell whitespace-nowrap max-w-[130px] h-full py-0 pl-0 lg:pl-16 pr-3 text-blueSilver font-bold text-base leading-5 tracking-wide">
+            <td className="hidden lg:table-cell whitespace-nowrap max-w-[130px] h-full py-0 pl-0 lg:pl-16 pr-3 font-bold text-base leading-5 tracking-wide">
               <div className="flex flex-col items-end gap-2 justify-end">
                 <span>{safetyScore}</span>
                 <SafetyScorePill score={safetyScore} />
@@ -128,7 +119,8 @@ const FarmsList: FC<Props> = ({ farms, positions }) => {
               )}
             </td>
             <td className="whitespace-nowrap max-w-[288px] py-4 pr-0 md:pr-6 xl:pr-14 text-right text-sm font-medium">
-              <div className="flex flex-row gap-x-3 items-center justify-start lg:justify-end">
+              <div className="flex flex-row gap-x-6 items-center justify-start lg:justify-end">
+                <ShareFarm farm={farm} />
                 <Button
                   size="large"
                   onButtonClick={() => {
@@ -137,9 +129,6 @@ const FarmsList: FC<Props> = ({ farms, positions }) => {
                 >
                   View Farm
                 </Button>
-                <div className="text-center scale-0 group-hover:scale-100 transition duration-200">
-                  <ShareFarm farm={farm} />
-                </div>
               </div>
             </td>
           </tr>
