+++ src/components/Library/Button.tsx
@@ -1,3 +1,4 @@
+import clsx from "clsx";
 import { ReactNode, memo } from "react";
 
 type ButtonProps = {
@@ -10,13 +11,15 @@ type ButtonProps = {
 const Button = ({ children, size, style, onButtonClick }: ButtonProps) => {
   return (
     <button
-      className={`flex flex-row items-center justify-center ring-2 text-base ring-[#314584] hover:ring-[#455b9c] text-white font-semibold rounded-xl leading-5 transition duration-200 ${style} ${
-        size == "small"
-          ? "py-2 px-[18px] sm:py-3 sm:px-6"
+      className={clsx(
+        "ring-1 bg-[#F9FAFB] ring-[#D0D5DD] text-[#1D2939] font-medium text-sm leading-5 rounded-lg transition duration-200 shadow-farmBtn",
+        style,
+        size == "large"
+          ? "py-[10px] sm:py-[14.5px] px-4"
           : size == "base"
           ? "py-[10.5px] px-6 rounded-[8px]"
-          : "py-[10.5px] sm:py-[14.5px] px-6"
-      }`}
+          : "py-2 px-[18px] sm:py-3 sm:px-6"
+      )}
       onClick={onButtonClick}
     >
       {children}
