+++ src/components/Library/ShareFarm.tsx
@@ -87,8 +87,8 @@ const ShareFarm = ({ farm }: ShareFarmPropsType) => {
 const ShareMenu = ({ farm, url, tweetUrl }: ShareMenuPropsType) => {
   return (
     <Menu as="div" className="relative hidden sm:inline-block">
-      <Menu.Button className="p-4 cursor-pointer transition-all duration-200">
-        <ShareIcon className="w-5 text-white" />
+      <Menu.Button className="p-3 cursor-pointer transition-all duration-200">
+        <ShareIcon className="w-6 text-[#AFAFAF]" />
       </Menu.Button>
 
       <Transition
