+++ tailwind.config.js
@@ -29,6 +29,7 @@ module.exports = {
       boxShadow: {
         tooltipLight: "0px 0px 5px #A2C4FF",
         tooltipDark: "0px 0px 5px #3780FF",
+        farmBtn: "0px 1px 2px rgba(16, 24, 40, 0.05)",
       },
       animation: {
         shiny: "shimmer 2s cubic-bezier(0.4, 0, 0.6, 1) infinite",
